+++ src/main/java/com/example/crawler/config/JpaConfig.java
@@ -1,4 +1,4 @@
-package com.example.webscraper.config;
+package com.example.crawler.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
