+++ src/main/java/com/example/crawler/meal/entity/MealItem.java
@@ -1,4 +1,4 @@
-package com.example.webscraper.meal.entity;
+package com.example.crawler.meal.entity;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
