+++ src/main/java/com/example/crawler/meal/service/MenuKafkaProducer.java
@@ -1,6 +1,6 @@
-package com.example.webscraper.meal.service;
+package com.example.crawler.meal.service;
 
-import com.example.webscraper.meal.entity.MealItem;
+import com.example.crawler.meal.entity.MealItem;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
